Implement job list with sqlc using templates (#795)
Here, reimplement the job list API's internals using sqlc using the new
sqlc template system introduced in #794. Admittedly this one isn't quite
as big of a win as we expect the templating to be in some cases, but it
gets rid of a case of SQL code in a raw string, and lets us drop the
`JobListFields` driver interface function.

We don't add tests because the APIs are relatively well tested already,
but make some small modifications to the driver `JobList` test cases so
they're a little more thorough.