More abstracted middleware modeled after hooks
Here, try to reuse some of the concepts developed for client "hooks"
that came in with #789 for middleware. The general premise is to make
hooks and middleware as close to each other as possible and to improve
the DX/syntax of using middleware.

A current DX problem with the middleware design is that for any
middleware that'd like to insert itself for multiple operations an
instance of it needs to go into both `JobInsertMiddleware` and
`WorkerMiddleware`, which generally means that it's initialized outside
of a `river.NewClient` invocation and making the code kind of ugly.
Here's an example of thatfrom an OpenTelemetry middleware I'm working on:

    middleware := riveropentelemetry.NewMiddleware(nil)

    _, err := river.NewClient(riverpgxv5.New(nil), &river.Config{
        JobInsertMiddleware: []rivertype.JobInsertMiddleware{
            middleware,
        },
        WorkerMiddleware: []rivertype.WorkerMiddleware{
            middleware,
        },
    })

Here, we re-integrate the two middleware stacks into one more abstracted
stack that enables a much more succinct injection:

    _, err := river.NewClient(riverpgxv5.New(nil), &river.Config{
        Middleware: []rivertype.Middleware{
            riveropentelemetry.NewMiddleware(nil),
        },
    })

Like with hooks, the middleware is internalized to a lookup struct so
that we precalculate what middleware belongs where and there's no
performance hit, if any.

The approach is a little less type safe in that it's possible to write a
trivial middleware that fulfills `rivertype.Middleware`, but accidentally
fails to implement a more useful middleware interface so that in the end
it doesn't actually run anything. We try to hedge against this by
suggesting that users write interface compliance checks like:

    // Verify interface compliance. It's recommended that these are included in your
    // test suite to make sure that your middlewares are complying to the specific
    // interface middlewares that you expected them to be.
    var (
        _ rivertype.JobInsertMiddleware = &JobInsertAndWorkMiddleware{}
        _ rivertype.WorkerMiddleware    = &JobInsertAndWorkMiddleware{}
        _ rivertype.JobInsertMiddleware = &JobInsertMiddleware{}
        _ rivertype.WorkerMiddleware    = &WorkerMiddleware{}
    )

Also take the opportunity to add some new example tests and tighten up
overly loose or add missing documentation in a few places.