Let `update-mod-version` take a `PACKAGE_PREFIX` env var
Here, have `update-mod-version` take a `PACKAGE_PREFIX` env var that
dictates which prefix it'll operate on, defaulting the previous behavior
of `github.com/riverqueue/river`.

The idea here is that we should be able to use reuse this command from
other River repositories like `rivercontrib` to update internal cross
dependencies.