Set minimum Go version to Go 1.23
As noted over in [1], it seems that `x/text` has already started
requiring Go 1.23 as its minimum version, meaning the writing's on the
wall for Go 1.22.

Go's official versioning policy is support for the last two minor
versions, and given Go 1.24's been out for some time now, it seems like
not a bad time to let 1.22 ride off into the sunset.

[1] https://github.com/riverqueue/riverui/pull/311#issuecomment-2730760360