Deprecate per-operation middleware operation defaults (#815)
As per #804, middleware has been restructed to look more like the design
of hooks, and I was finding that as I restructuring documentation as
part of [1] it was hard to recommend the use of per-operation middleware
defaults.

Here, deprecate the per-operation defaults in favor of just the single
`river.MiddlewareDefaults` struct. This should also be somewhat more
beneficial in case more middleware operations need to be added in the
future because we'll just have fewer total new types emerge out of it.

[1] https://github.com/riverqueue/homepage/pull/198