Upgrade golangci-lint to v2
For better or rwose, golangci-lint has released a v2 [1] that's
incompatible with v1. It'll only be a matter of time until we're forced
to upgrade to it, so I figure we may as well do it early.

There's a bunch of new obnoxious linters that I've disabled. Most of
these complain about "cyclomatic complexity" or "cognitive complexity"
or functions being too long (they all seem to want everything broken
down into functions of ten lines).

Aside from that, `testifylint` seems to have gotten even more effective
at finding misuses of testify, so it fixed a bunch more arguable misuses
of various assertions, which seems good for normalizing code.

The structure of the YAML file changes fairly significantly, but it's
all basically the same thing with various aspects reorganized. I made
the new one using the built-in `golangci-lint migrate` and copied
comments over manually.

If we can get this in, the next step will be to distribute it out to
other River projects. Most of them are using copies of this repo's
configuration file, so it should be pretty easy to do.

[1] https://ldez.github.io/blog/2025/03/23/golangci-lint-v2/