Replace uses of `valutil.FirstNonZero` with `cmp.Or` + deprecate former (#845)
`cmp.Or` wasn't available originally in River, but was add in 1.22.
We're now minimum version 1.23, so it's fine to switch over it given it
does exactly the same thing as the custom `valutil.FirstNonZero`.

We have some dependencies using `valutil.FirstNonZero` so we shouldn't
remove it until there's at least one more version of those that've been
cut to avoid potential version compatibility problems. For now, mark the
function as deprecated to warn against use.