Use schema testing for `TestTx`, sharing one schema per set of migration lines
Here, start using a new isolated schema for test transaction invocations
rather than falling back on whatever happens to be in `river_test`. A
major benefit of this approach is that it makes migration inconsistencies
impossible so that when checking out a branch with a new migration, your
tests will always get the new migration, but then drop it when moving
back to the `master` branch.

It has the additional advantage that when alternatively running River
and River Pro tests locally pointing to the same `river_test` database,
we get the right migration lines raised for `TestTx` without having to
worry about whether Pro migrations have been made available yet or not.

We attempt to keep test transactions very efficient by only raising one
schema per package being tested (and per set of migration lines,
although this is rarely used) so we don't need to create a schema and
run migrations once per test, but rather once for _all_ tests in the
package.

Similar to `TestSchema`, `TestTx` picks up a driver so that it can use
`GetMigrationDefaultLines` to determine what migration lines a test case
is expecting to exist before running. Once again, this is mainly useful
for River Pro.

`TestTx` gets moved to the higher level `riverschematest` package
because it needs access to high level constructs like migrations, and I
rename `riverschematest` to `riverdbtest` because with test transactions
in it, its use becomes more general. I would've also moved functions
in `riversharedtest` like `DBPool` and `TestDatabaseURL` over, but
unfortunately we still need these at a lower level so that the tests in
`rivermigrate` can use them.

I removed the `riverdbtest` functionality that checks for extraneous
rows. The `public` schema no longer even exists in CI, and the check is
fairly unnecessary now that zeroing `search_path` makes it very
difficult to accidentally insert records into anything but the current
test schema.